feat(lobby): change user_id type from int64 to uint32
- Update checkin/store.go: change userID parameter type from int64 to uint32
- Update item/store.go: change userID parameter type from int64 to uint32
- Update checkin_server.go: change ctx.Value("user_id") assertion to uint32
- Update item_server.go: change ctx.Value("user_id") assertion to uint32
- Update all test files to use uint32 for user_id context values

This aligns with the proto change from uint64 to uint32 for consistency.

+++ services/lobby/cmd/lobby/checkin_server.go
@@ -56,7 +56,7 @@ func (s *CheckinService) SetContinuousRewards(rewards map[int]map[int32]int64) {
 
 // Checkin 签到
 func (s *CheckinService) Checkin(ctx context.Context, req *lobbyv1.CheckinRequest) (*lobbyv1.CheckinData, error) {
-	userID, ok := ctx.Value("user_id").(int64)
+	userID, ok := ctx.Value("user_id").(uint32)
 	if !ok {
 		return nil, status.Errorf(codes.Code(commonv1.ErrorCode_USER_NOT_AUTHENTICATED), "用户未认证")
 	}
@@ -128,7 +128,7 @@ func (s *CheckinService) Checkin(ctx context.Context, req *lobbyv1.CheckinReques
 
 // GetStatus 获取签到状态
 func (s *CheckinService) GetStatus(ctx context.Context, req *lobbyv1.GetStatusRequest) (*lobbyv1.CheckinStatusData, error) {
-	userID, ok := ctx.Value("user_id").(int64)
+	userID, ok := ctx.Value("user_id").(uint32)
 	if !ok {
 		return nil, status.Errorf(codes.Code(commonv1.ErrorCode_USER_NOT_AUTHENTICATED), "用户未认证")
 	}
