+++ services/lobby/cmd/lobby/item_server.go
@@ -28,7 +28,7 @@ func NewItemService(rdb *redis.Client) *ItemService {
 // GetMyItems 获取我的物品
 func (s *ItemService) GetMyItems(ctx context.Context, req *lobbyv1.GetMyItemsRequest) (*lobbyv1.MyItemsData, error) {
 	// 从 context 获取 user_id（由 Gateway 注入）
-	userID, ok := ctx.Value("user_id").(int64)
+	userID, ok := ctx.Value("user_id").(uint32)
 	if !ok {
 		return nil, status.Errorf(codes.Code(commonv1.ErrorCode_USER_NOT_AUTHENTICATED), "用户未认证")
 	}
